refactor: movable 패키지 생성 및 코드 리팩토링
movable 패키지 생성, 기존 Generatenumber 패키지 역할 수행
ALL 구조 변경에 맞춘 코드 리팩토링
