+++ src/test/java/racinggame/domain/RacingGameTest.java
@@ -5,8 +5,6 @@ import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import racinggame.domain.car.Cars;
-import racinggame.domain.car.generatenumber.MovAbleStrategy;
-import racinggame.domain.car.generatenumber.RandomMovableStrategy;
 import racinggame.domain.data.Names;
 import racinggame.domain.data.GameStatus;
 import racinggame.domain.car.Car;
